+++ src/octaves.rs
@@ -5,6 +5,12 @@ use num_traits::Zero;
 use higher_order_functions::Init;
 
 /// A noise function formed by summing several octaves of another noise function.
+///
+/// `lacunarity`: A multiplier for the frequency from one octave to the next.
+///
+/// `persistence`: A multiplier for the persistence from one octave to the next.
+///
+/// A good default is `Octaves::new(_, 0.5, 0.5)`.
 #[derive(Copy, Clone)]
 pub struct Octaves<Inner, const N: usize> {
 	inner: SumNoise<ScaleNoise<Inner, f64, f64>, N>,
@@ -19,13 +25,6 @@ impl<Inner: Noise, const N: usize> Noise for Octaves<Inner, N> where
 }
 
 impl<Inner: Copy, const N: usize> Octaves<Inner, N> {
-	/// Create a noise function by summing several octaves of another noise function.
-	///
-	/// `lacunarity`: A multiplier for the frequency from one octave to the next.
-	///
-	/// `persistence`: A multiplier for the persistence from one octave to the next.
-	///
-	/// A good default is `Octaves::new(_, 0.5, 0.5)`.
 	pub fn new(inner: Inner, lacunarity: f64, persistence: f64) -> Octaves<Inner, N> {
 		Octaves { inner: SumNoise::new(<[_; N]>::init(|i| ScaleNoise::new(inner, lacunarity.powi(i as i32), persistence.powi(i as i32)))) }
 	}