+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.1.1](https://crates.io/crates/noise_fn/0.1.1) - 2020-09-11
+
+### Added:
+* `AddNoise`, `Constant`, and `Gradient` noise functions.
+* `IgnoreSeed`, a workaround to allow seedable and non-seedable noise functions to be used together.
+
 ## [v0.1.0](https://crates.io/crates/noise_fn/0.1.0) - 2020-09-11
 
 ### Added: